Move test-specific stuff to *Utils objects
