+++ zio-config/src/test/scala/zio/config/ReadErrorsTest.scala
@@ -1,6 +1,8 @@
 package zio.config
 
-import zio.config.helpers._
+import zio.config.ReadError.ParseError
+import zio.config.ReadErrorsTestUtils._
+import zio.random.Random
 import zio.test._
 import zio.test.Assertion._
 
@@ -15,3 +17,22 @@ object ReadErrorsTest
         }
       )
     )
+
+object ReadErrorsTestUtils {
+  private val genParseError =
+    for {
+      s1 <- Gen.anyString
+      s2 <- Gen.anyString
+      s3 <- Gen.anyString
+    } yield ParseError(s1, s2, s3)
+
+  private val genReadError =
+    Gen.oneOf(Gen.const(ReadError.MissingValue("somekey")), genParseError)
+
+  val genReadErrors: Gen[Random with Sized, List[ReadError]] = {
+    for {
+      n    <- Gen.int(1, 20)
+      list <- Gen.listOfN(n)(genReadError)
+    } yield list
+  }
+}
